pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "llm-cost-estimator"
version     = "0.1.0"
description = "CLI tool to estimate LLM token counts and API costs"
readme      = "README.md"
requires-python = ">=3.9"

dependencies = [
    "tiktoken>=0.7.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
llm-cost = "llm_cost.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_cost*"]